Rename BindMode namespace to Types and update references throughout the codebase
